Full_app
Разбил по модулым: start, interace, input, rational_calc, out
При правильном вводе через пробел программа отрабатывает и выводит дробь.
