+++ start.py
@@ -1,0 +1,7 @@
+from interface import start_inteface
+
+def main_rationals():
+    start_inteface()
+
+if __name__ == "__main__":
+    main_rationals()
